add logout feature
