+++ app_server/api/urls.py
@@ -1,7 +1,8 @@
 from django.urls import path
 from . import views
 from api.views.personalInfo import PersonalInfoView,PersonalInfoModifyView
-from api.views import register
+from api.views.test_auth import WelcomeAuth
+from api.views.authentication import signup,logout
 from rest_framework.authtoken.views import obtain_auth_token
 
 urlpatterns = [
@@ -9,7 +10,11 @@ urlpatterns = [
     #path('personalInformations/<int:id>/', PersonalInformationsView.as_view()),
     #path('personalInformations/modify/<int:id>/', PersonalInformationsModifyView.as_view())
     path('login/', obtain_auth_token, name='api_token_auth'),  # <-- And here
-    path('signup/', register.signup, name="signup"),
+    path('logout/', logout, name="logout"),
+    path('signup/', signup, name="signup"),
+    path('welcomeAuth/', WelcomeAuth.as_view()),
+
+
     path('personalInfo/<int:id>/', PersonalInfoView.as_view()),
     path('personalInfo/', PersonalInfoView.as_view()),
     path('personalInfo/modify/', PersonalInfoModifyView.as_view())
